Added Sandbox scene, cleaned up scripts
