+++ Assets/SendGameInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e1d686a61d248b4a9c9b755036f12c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
