+++ README.md
@@ -1,0 +1,8 @@
+# Distributed Shooter
+
+## Distributed Systems course - Programming Task 2021
+
+Project [documentation](https://drive.google.com/drive/u/0/folders/1YdaHjym6uTl3Xfy7XMbDgkyaTkJP7su9)
+
+Project [website](https://meldv.github.io/UnityGameData/)
+Website [repo](https://github.com/MelDv/UnityGameData)
