+++ UserSettings/EditorUserSettings.asset
@@ -15,18 +15,21 @@ EditorUserSettings:
       value: 56075155035408590e5b0e20437359444f164e722e2a7e367c7b4461e4b96c60
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 5105515f000059020e5e552040715b444215417974787231757b1b6bbae66c60
-      flags: 0
-    RecentlyUsedSceneGuid-4:
       value: 5403515306035c5a595b547b43740e44424f4c7a7c7a7f61792c4e35e6b56669
       flags: 0
-    RecentlyUsedSceneGuid-5:
+    RecentlyUsedSceneGuid-4:
       value: 5302030255030f59080a097012770f444716407e7f2c7f35287f1b36b5b5643b
       flags: 0
+    RecentlyUsedSceneGuid-5:
+      value: 5105515f000059020e5e552040715b444215417974787231757b1b6bbae66c60
+      flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 0601055e00055859580c5b70417a5a4440161d7a7e7e73322b2b4e65e7e4633d
+      value: 5753575504575e5e0b5a087246205d4414164a28297c72647e7e1c6be0b33760
       flags: 0
     RecentlyUsedSceneGuid-7:
+      value: 0601055e00055859580c5b70417a5a4440161d7a7e7e73322b2b4e65e7e4633d
+      flags: 0
+    RecentlyUsedSceneGuid-8:
       value: 52570d5e03050f0858565b2415255b4441151d287d2977692e7f4a6abbb66668
       flags: 0
     RecentlyUsedScenePath-0:
